Remove course application as there can be only one application for app
- Since android app should have only one application class, sdk should not
have application as main app will have it.
- Code in application is moved to appropriate classes
